GN-168: Reorganised Metrics
