+++ Domain.Supply/Entities/Supply/FeedingSupply/FeedingSupplyNutritionEntity.cs
@@ -5,7 +5,7 @@ namespace Diagnosea.Submarine.Domain.Supply.Entities.Supply.FeedingSupply
     public class FeedingSupplyNutritionEntity
     {
         public string Name { get; set; }
-        public int Amount { get; set; }
-        public Measurement Measurement { get; set; }
+        public int Measurement { get; set; }
+        public Metric Metric { get; set; }
     }
 }
