+++ Domain.Supply/Entities/Supply/MedicalSupplyEntity.cs
@@ -1,9 +1,16 @@
-﻿using Diagnosea.Submarine.Abstractions.Enums;
+﻿using System.Collections.Generic;
+using Diagnosea.Submarine.Abstractions.Enums;
+using Diagnosea.Submarine.Abstractions.Enums.Supply;
 
 namespace Diagnosea.Submarine.Domain.Supply.Entities.Supply
 {
     public class MedicalSupplyEntity : SupplyEntity
     {
-        public Ailment Ailment { get; set; }
+        public IList<Ailment> Treats { get; set; }
+
+        public MedicalSupplyEntity()
+        {
+            Category = SupplyCategory.Medical;
+        }
     }
 }
