+++ Domain.Supply/Entities/Supply/SupplyDurability/SupplyDurabilityEntity.cs
@@ -1,0 +1,9 @@
+﻿using Diagnosea.Submarine.Abstractions.Enums;
+
+namespace Diagnosea.Submarine.Domain.Supply.Entities.Supply.SupplyDurability
+{
+    public class SupplyDurabilityEntity {
+        public int Measurement { get; set; }
+        public Metric Metric { get; set; }
+    }
+}
