+++ Domain.Supply/Entities/Supply/SupplyEntity.cs
@@ -1,5 +1,10 @@
 ﻿using System;
+using System.Collections.Generic;
 using Diagnosea.Submarine.Abstractions.Enums.Supply;
+using Diagnosea.Submarine.Domain.Supply.Entities.Supply.SupplyCompatibility;
+using Diagnosea.Submarine.Domain.Supply.Entities.Supply.SupplyDurability;
+using Diagnosea.Submarine.Domain.Supply.Entities.Supply.SupplyQuantity;
+using Diagnosea.Submarine.Domain.Supply.Entities.Supply.SupplyRequirement;
 
 namespace Diagnosea.Submarine.Domain.Supply.Entities.Supply
 {
@@ -9,6 +14,9 @@ namespace Diagnosea.Submarine.Domain.Supply.Entities.Supply
         public Guid SupplyStoreId { get; set; }
         public string Name { get; set; }
         public SupplyCategory Category { get; set; }
-        public int OwnedQuantity { get; set; }
+        public SupplyDurabilityEntity Durability { get; set; }
+        public SupplyQuantityEntity Owned { get; set; }
+        public IList<SupplyCompatibilityEntity> Compatibilities { get; set; }
+        public IList<SupplyRequirementEntity> Requirements { get; set; }
     }
 }
