+++ Domain.Supply/Entities/Supply/SupplyQuantity/SupplyQuantityEntity.cs
@@ -1,0 +1,10 @@
+﻿using Diagnosea.Submarine.Abstractions.Enums;
+
+namespace Diagnosea.Submarine.Domain.Supply.Entities.Supply.SupplyQuantity
+{
+    public class SupplyQuantityEntity
+    {
+        public int Measurement { get; set; }
+        public Metric Metric { get; set; }
+    }
+}
