--- Submarine Abstractions/Abstractions/Enums/Measurement.cs
@@ -1,7 +1,0 @@
-﻿namespace Diagnosea.Submarine.Abstractions.Enums
-{
-    public enum Measurement
-    {
-        Grams
-    }
-}
